Add an option to krenew to keep running despite errors
Add the -i option to krenew, which says to keep running even if there
is an error renewing the ticket cache.  This is useful if the ticket
cache renewed by krenew may expire and then later be renewed (such as
with a manual kinit) and krenew is expected to wake up again and
process the new ticket cache.