Rewrite Kerberos v4 configure probes
Rewrite the configure probes for Kerberos v4 to use the same style as
the Kerberos v5 configure probes.  Among the improvements are falling
back on direct library probing if the results of krb5-config fail, the
addition of --with-krb4-include and --with-krb4-lib flags, and support
for probing for optional Kerberos v4 libraries.