Update kafs replacement code and Autoconf probes from rra-c-util
The kafs code is now in a kafs directory instead of in portable and uses
an Automake conditional instead of AC_LIBOBJ to add it to the compilation.
The Autoconf code now follows my normal Autoconf coding style, including
separate --with-*-include and --with-*-lib configure options for libkafs
and the AFS libraries (if needed).  The kafs replacement also now has a
test suite.