Allow the argument to -k to start with FILE:
Allow the argument to -k to start with FILE: and strip off that prefix
to form the ticket cache name.  -k still forces its argument to be a
file-based cache, however; FILE: is the only cache type designator
supported.  Use the KRB5CCNAME environment variable for other ticket
cache types.