Rewrite krenew to use a new generic framework
The new framework will be used for both k5start and krenew and
refactors the common code between them, including the infrastructure
for running commands and running as a daemon.

Clean up handling of when to exit on errors.  krenew now defaults
to remaining running if there's an error in contacting the KDC or
storing the new credentials, only exiting by default if the renewable
lifetime has expired or if the ticket cache has disappeared.  -i can
still be used to keep it running in those cases as well.  However,
without -i, it will always exit if the initial authentication fails.

krenew now removes any PID files that it writes on exit, and is better
about removing a copied ticket cache on exit.