k5start -K no longer exits if initial authentication fails
k5start, when run with the -K option to run as a daemon, no longer
exits if the initial authentication fails.  Instead, it reports the
error to standard error and then continues to run, attempting
authentication every minute as if authentication had failed after it
had started.  Patch from Rasmus Borup Hansen.