Fix races in k5start/daemon and krenew/daemon tests
Several of the tests for backgrounding were missing a delay to
wait for the daemon to start.  Add those in, and clean up the
PID files from previous tests to avoid spurious failures.

k5start and krenew do not write their PID file atomically, so
there was a race condition in the tests where they would detect
the existence of the file and then try to read an empty file.
Avoid this by checking that the file contains a PID using -s.