Fix k5start/perms test on Heimdal
The Heimdal klist, even when it thinks it's running as root,
refuses to open a ticket cache owned by another user or with a
publicly-readable mode.  Work around this by checking the mode
and ownership first, then resetting them to what was expected
and checking the contents of the cache.